darek_web/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
Pillow==10.4.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
twilio==9.8.5
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
